refactor: migrate save_profiler
